Changes to variable types to be cleaner and more efficient in 64-bit mode and protect against underflow, especially for array indexes and format specifiers.
git-svn-id: http://dysart.cs.byu.edu/chsvn/trunk@531 a75b6edf-6070-4da8-89c8-9992cc802e51

+++ test/CHCustomDictionariesTest.m
@@ -45,7 +45,7 @@
 		if (allKeys == nil)
 			allKeys = [dictionary allKeys];
 		NSUInteger count = MIN([dictionary count], [expectedKeyOrder count]);
-		for (int i = 0; i < count; i++) {
+		for (NSUInteger i = 0; i < count; i++) {
 			STAssertEqualObjects([allKeys objectAtIndex:i],
 								 [expectedKeyOrder objectAtIndex:i],
 								 @"Wrong output ordering of keys.");
